[WIP][package] Factor out dependency resolution
**Summary**
This commit separates the dependency resolution logic in
`PackageExporter` out into a separate class. `PackageExporter` contains
the logic for saving/pickling. The motivation for this is to reuse
dependency resolution code as the basis for dependency exploration and
visualization tools.

[ghstack-poisoned]
